move stable_narwhalify_test to v1_test
